tests: separate docs tests
