+++ docs/docs/SUMMARY.md
@@ -123,23 +123,6 @@ search:
             - [apply_types](api/faststream/apply_types.md)
             - app
                 - [FastStream](api/faststream/app/FastStream.md)
-            - asgi
-                - [AsgiFastStream](api/faststream/asgi/AsgiFastStream.md)
-                - [AsgiResponse](api/faststream/asgi/AsgiResponse.md)
-                - [get](api/faststream/asgi/get.md)
-                - [make_asyncapi_asgi](api/faststream/asgi/make_asyncapi_asgi.md)
-                - [make_ping_asgi](api/faststream/asgi/make_ping_asgi.md)
-                - app
-                    - [AsgiFastStream](api/faststream/asgi/app/AsgiFastStream.md)
-                - factories
-                    - [make_asyncapi_asgi](api/faststream/asgi/factories/make_asyncapi_asgi.md)
-                    - [make_ping_asgi](api/faststream/asgi/factories/make_ping_asgi.md)
-                - handlers
-                    - [get](api/faststream/asgi/handlers/get.md)
-                - response
-                    - [AsgiResponse](api/faststream/asgi/response/AsgiResponse.md)
-                - websocket
-                    - [WebSocketClose](api/faststream/asgi/websocket/WebSocketClose.md)
             - asyncapi
                 - [get_app_schema](api/faststream/asyncapi/get_app_schema.md)
                 - [get_asyncapi_html](api/faststream/asyncapi/get_asyncapi_html.md)
