--- docs/docs/en/api/faststream/asgi/websocket/WebSocketClose.md
@@ -1,11 +1,0 @@
----
-# 0.5 - API
-# 2 - Release
-# 3 - Contributing
-# 5 - Template Page
-# 10 - Default
-search:
-  boost: 0.5
----
-
-::: faststream.asgi.websocket.WebSocketClose
